Refactor graphql client to be more modular
