--- src/graphql/subscriptions.ts
@@ -1,6 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-import { gql } from '@apollo/client';
-
-/*
- * Define graphql subscriptions here and then `@graphql-codegen` can generate TypeScript code for those.
- */
